refactor(dto): extract task IPC contracts
- move Task request and response types into the DTO layer

- keep Task command handlers focused on transport orchestration
