+++ src-tauri/src/lib.rs
@@ -28,6 +28,7 @@ mod dto {
     pub(crate) mod agent;
     pub(crate) mod comparison;
     pub(crate) mod skill;
+    pub(crate) mod task;
     pub(crate) mod workspace;
 }
 mod domain {
